+++ frontend/src/components/standings/GroupTable.tsx
@@ -7,7 +7,7 @@ import {
   TableRow,
 } from '@/components/ui/table';
 
-import { type Standing } from '@/types/standings';
+import { type Standing } from '@/types/standing';
 
 import { COLUMNS } from '@/constants/standings';
